move encounter form to reusable component
